actually use the server struct
